+++ src/components/App.spec.js
@@ -1,4 +1,6 @@
+
 import App from './App.js';
+
 
 describe('App', () => {
   it('should render without crashing', () => {
@@ -6,3 +8,5 @@ describe('App', () => {
     expect(el instanceof HTMLElement).toBe(true);
   });
 });
+
+
