+++ src/index.html
@@ -9,10 +9,11 @@
       content="width=device-width, user-scalable=no, initial-scale=1.0, maximun-scale=1.0, minimun.scale=1.0  " />
    <a href="https://fontmeme.com/es/fuente-pokemon/"></a>
    <link rel="stylesheet" media="only screen and (max-width: 768px)" href="style.css">
+   <script src="main.js" type="module"></script>
 </head>
 
 <body id="root">
-   <div id="pantalla1" style="display: block;">
+   <div id="screenOne" style="display: block;"><!--Aqui comienza panttalla uno -->
       <div id="containerStepOne" class="container">
          <div class="boxImg">
             <img class="img" src="/assets/pokemon.png" />
@@ -30,33 +31,18 @@
 
          <div class="boxButton">
             <img class="imgTwo" src="/assets/clipart967980.png" />
-            <button id="buttonStart" class="redondo"><img class="imgBoton" src="/assets/botonpoke.png">
+            <button id="buttonStart" class="buttonStyle"><img class="imgBoton" src="/assets/botonpoke.png">
                <p> <img class="pointer" src="/assets/click.png"></p>
             </button>
          </div>
-         <script src="main.js" type="module"></script>
-      </div>
-   </div>
-   <div id="pantalla2" style="display: none;">
-      <div id="containerStepTwo" class="containerFlextwo">
-         <div class="timer">
-            <div class="baseTimer">
-               <svg class="baseTimerSVG" viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
-                  <g class="baseTimerCircle">
-                     <circle class="baseTimerPathElapsed" cx="50" cy="50" r="45"></circle>
-                     <path id="baseTimerPathRemaining" stroke-dasharray="283"
-                        class="baseTimerPathRemaining ${remainingPathColor}" d="
-                       M 50, 50
-                       m -45, 0
-                       a 45,45 0 1,0 90,0
-                       a 45,45 0 1,0 -90,0
-                     "></path>
-                  </g>
-               </svg>
-               <span id="baseTimerLabel" class="baseTimerLabelONE">
 
-               </span>
-            </div>
+      </div>
+   </div> <!-- Aqui termina panttalla uno -->
+
+   <div id="screenTwo" style="display: none;"> <!-- Aqui empieza panttalla dos -->
+
+      <div id="containerStepTwo" class="containerFlextwo">
+        <div class="timer">timer
          </div>
 
 
@@ -73,7 +59,7 @@
             <p class="boxWordhowToplay">
 
                <img src="https://fontmeme.com/permalink/210612/3196bc2870d25d7f1d446fc2904f3c3a.png"
-                  class="palabraComoJugar">
+                  class="boxWordhowToplayOne">
                <b>
 
                   <img class="pokemonIn" src="/assets/008.png"><img class="plus" src="/assets/+2.png"><img class="pokemonIn"
@@ -84,44 +70,47 @@
                   <img class="plus" src="/assets/+.png">
                   <img class="pokemonIn" src="/assets/WARTORTLE.png">
                   <img class="iqual" src="/assets/=.png">
-                  <img class="equis" src="/assets/aversion.png" class="equis">
+                  <img class="thumbsDown" src="/assets/aversion.png">
                </b>
 
 
             </p>
          </div>
       </div>
-   </div>
-   <div id="pantalla3" style="display: none;">
+   </div> <!-- Aqui termina panttalla dos -->
+
+   <div id="screenThree" style="display: none;"> <!-- Aqui empieza panttalla tres -->
+
    <div id="containerThree" class="containerFlexThree">
       <div>
-         <div class="cajaGanaste">
-         <img class="imgGanaste" src="assets/Ganaste.png">
+         <div class="boxWin">
+         <img class="imgWin" src="assets/Ganaste.png">
       </div>
 
-         <div class="CajaMeow">
+         <div class="boxMeow">
          <img class="meow" src="assets/meow.png">
       </div>
 
-         <div class="EquipoRRRR">
-         <img class="equipoRo" src="assets/equipo rocket.png">
+         <div class="teamRocket">
+         <img class="teamRocketOne" src="assets/equipo rocket.png">
       </div>
 
       <div class="boxR">
          <img class="r" src="assets/r.png">
       </div>
 
-         <div class="cajaPikachuFeliz">
-         <img class="pikachuFeliz" src="assets/pikachu feliz.png">
+         <div class="pikachuBoxtwo">
+         <img class="smilingPikachu" src="assets/pikachu feliz.png">
       </div>
 
          <div class="boxButtonTwo">
-         <button id="reset" class="button" type="reset"><img class="volveraJugar" src="assets/volver a jugar.png">
+         <button id="reset" class="button" type="reset"><img class="playAgain" src="assets/volver a jugar.png">
          </button>
       </div>
       </div>
    </div>
-   </div>
+   </div> <!-- Aqui termina panttalla tres -->
+
 </body>
 
 <footer class="footer">Derechos reservados @Nintendo</footer>
